+++ components/Chips/src/MDCChipField.m
@@ -428,7 +428,7 @@ const UIEdgeInsets MDCChipFieldDefaultContentEdgeInsets = {
   CGFloat padding =
       (MDCChipFieldClearButtonSquareWidthHeight - MDCChipFieldClearImageSquareWidthHeight) / 2;
   clearImageView.frame = CGRectMake(padding, padding, widthAndHeight, widthAndHeight);
-  clearButton.tintColor = [UIColor.blackColor colorWithAlphaComponent:0.6f];
+  clearButton.tintColor = [UIColor.blackColor colorWithAlphaComponent:(CGFloat)0.6];
   [clearButton addSubview:clearImageView];
   chip.accessoryView = clearButton;
   [clearButton addTarget:self
@@ -455,80 +455,80 @@ static inline UIBezierPath *MDCPathForClearButtonImageFrame(CGRect frame) {
   // GENERATED CODE
 
   CGRect innerBounds = CGRectMake(CGRectGetMinX(frame) + 2, CGRectGetMinY(frame) + 2,
-                                  MDCFloor((frame.size.width - 2) * 0.90909f + 0.5f),
-                                  MDCFloor((frame.size.height - 2) * 0.90909f + 0.5f));
+                                  MDCFloor((frame.size.width - 2) * (CGFloat)0.90909 + (CGFloat)0.5),
+                                  MDCFloor((frame.size.height - 2) * (CGFloat)0.90909 + (CGFloat)0.5));
   UIBezierPath *ic_clear_path = [UIBezierPath bezierPath];
   [ic_clear_path
-      moveToPoint:CGPointMake(CGRectGetMinX(innerBounds) + 0.50000f * innerBounds.size.width,
+      moveToPoint:CGPointMake(CGRectGetMinX(innerBounds) + (CGFloat)0.50000 * innerBounds.size.width,
                               CGRectGetMinY(innerBounds) + 0 * innerBounds.size.height)];
   [ic_clear_path
       addCurveToPoint:CGPointMake(CGRectGetMinX(innerBounds) + 1 * innerBounds.size.width,
-                                  CGRectGetMinY(innerBounds) + 0.50000f * innerBounds.size.height)
-        controlPoint1:CGPointMake(CGRectGetMinX(innerBounds) + 0.77600f * innerBounds.size.width,
+                                  CGRectGetMinY(innerBounds) + (CGFloat)0.50000 * innerBounds.size.height)
+        controlPoint1:CGPointMake(CGRectGetMinX(innerBounds) + (CGFloat)0.77600 * innerBounds.size.width,
                                   CGRectGetMinY(innerBounds) + 0 * innerBounds.size.height)
         controlPoint2:CGPointMake(CGRectGetMinX(innerBounds) + 1 * innerBounds.size.width,
-                                  CGRectGetMinY(innerBounds) + 0.22400f * innerBounds.size.height)];
+                                  CGRectGetMinY(innerBounds) + (CGFloat)0.22400 * innerBounds.size.height)];
   [ic_clear_path
-      addCurveToPoint:CGPointMake(CGRectGetMinX(innerBounds) + 0.50000f * innerBounds.size.width,
+      addCurveToPoint:CGPointMake(CGRectGetMinX(innerBounds) + (CGFloat)0.50000 * innerBounds.size.width,
                                   CGRectGetMinY(innerBounds) + 1 * innerBounds.size.height)
         controlPoint1:CGPointMake(CGRectGetMinX(innerBounds) + 1 * innerBounds.size.width,
-                                  CGRectGetMinY(innerBounds) + 0.77600f * innerBounds.size.height)
-        controlPoint2:CGPointMake(CGRectGetMinX(innerBounds) + 0.77600f * innerBounds.size.width,
+                                  CGRectGetMinY(innerBounds) + (CGFloat)0.77600 * innerBounds.size.height)
+        controlPoint2:CGPointMake(CGRectGetMinX(innerBounds) + (CGFloat)0.77600 * innerBounds.size.width,
                                   CGRectGetMinY(innerBounds) + 1 * innerBounds.size.height)];
   [ic_clear_path
       addCurveToPoint:CGPointMake(CGRectGetMinX(innerBounds) + 0 * innerBounds.size.width,
-                                  CGRectGetMinY(innerBounds) + 0.50000f * innerBounds.size.height)
-        controlPoint1:CGPointMake(CGRectGetMinX(innerBounds) + 0.22400f * innerBounds.size.width,
+                                  CGRectGetMinY(innerBounds) + (CGFloat)0.50000 * innerBounds.size.height)
+        controlPoint1:CGPointMake(CGRectGetMinX(innerBounds) + (CGFloat)0.22400 * innerBounds.size.width,
                                   CGRectGetMinY(innerBounds) + 1 * innerBounds.size.height)
         controlPoint2:CGPointMake(CGRectGetMinX(innerBounds) + 0 * innerBounds.size.width,
-                                  CGRectGetMinY(innerBounds) + 0.77600f * innerBounds.size.height)];
+                                  CGRectGetMinY(innerBounds) + (CGFloat)0.77600 * innerBounds.size.height)];
   [ic_clear_path
-      addCurveToPoint:CGPointMake(CGRectGetMinX(innerBounds) + 0.50000f * innerBounds.size.width,
+      addCurveToPoint:CGPointMake(CGRectGetMinX(innerBounds) + (CGFloat)0.50000 * innerBounds.size.width,
                                   CGRectGetMinY(innerBounds) + 0 * innerBounds.size.height)
         controlPoint1:CGPointMake(CGRectGetMinX(innerBounds) + 0 * innerBounds.size.width,
-                                  CGRectGetMinY(innerBounds) + 0.22400f * innerBounds.size.height)
-        controlPoint2:CGPointMake(CGRectGetMinX(innerBounds) + 0.22400f * innerBounds.size.width,
+                                  CGRectGetMinY(innerBounds) + (CGFloat)0.22400 * innerBounds.size.height)
+        controlPoint2:CGPointMake(CGRectGetMinX(innerBounds) + (CGFloat)0.22400 * innerBounds.size.width,
                                   CGRectGetMinY(innerBounds) + 0 * innerBounds.size.height)];
   [ic_clear_path closePath];
   [ic_clear_path
-      moveToPoint:CGPointMake(CGRectGetMinX(innerBounds) + 0.73417f * innerBounds.size.width,
-                              CGRectGetMinY(innerBounds) + 0.31467f * innerBounds.size.height)];
+      moveToPoint:CGPointMake(CGRectGetMinX(innerBounds) + (CGFloat)0.73417 * innerBounds.size.width,
+                              CGRectGetMinY(innerBounds) + (CGFloat)0.31467 * innerBounds.size.height)];
   [ic_clear_path
-      addLineToPoint:CGPointMake(CGRectGetMinX(innerBounds) + 0.68700f * innerBounds.size.width,
-                                 CGRectGetMinY(innerBounds) + 0.26750f * innerBounds.size.height)];
+      addLineToPoint:CGPointMake(CGRectGetMinX(innerBounds) + (CGFloat)0.68700 * innerBounds.size.width,
+                                 CGRectGetMinY(innerBounds) + (CGFloat)0.26750 * innerBounds.size.height)];
   [ic_clear_path
-      addLineToPoint:CGPointMake(CGRectGetMinX(innerBounds) + 0.50083f * innerBounds.size.width,
-                                 CGRectGetMinY(innerBounds) + 0.45367f * innerBounds.size.height)];
+      addLineToPoint:CGPointMake(CGRectGetMinX(innerBounds) + (CGFloat)0.50083 * innerBounds.size.width,
+                                 CGRectGetMinY(innerBounds) + (CGFloat)0.45367 * innerBounds.size.height)];
   [ic_clear_path
-      addLineToPoint:CGPointMake(CGRectGetMinX(innerBounds) + 0.31467f * innerBounds.size.width,
-                                 CGRectGetMinY(innerBounds) + 0.26750f * innerBounds.size.height)];
+      addLineToPoint:CGPointMake(CGRectGetMinX(innerBounds) + (CGFloat)0.31467 * innerBounds.size.width,
+                                 CGRectGetMinY(innerBounds) + (CGFloat)0.26750 * innerBounds.size.height)];
   [ic_clear_path
-      addLineToPoint:CGPointMake(CGRectGetMinX(innerBounds) + 0.26750f * innerBounds.size.width,
-                                 CGRectGetMinY(innerBounds) + 0.31467f * innerBounds.size.height)];
+      addLineToPoint:CGPointMake(CGRectGetMinX(innerBounds) + (CGFloat)0.26750 * innerBounds.size.width,
+                                 CGRectGetMinY(innerBounds) + (CGFloat)0.31467 * innerBounds.size.height)];
   [ic_clear_path
-      addLineToPoint:CGPointMake(CGRectGetMinX(innerBounds) + 0.45367f * innerBounds.size.width,
-                                 CGRectGetMinY(innerBounds) + 0.50083f * innerBounds.size.height)];
+      addLineToPoint:CGPointMake(CGRectGetMinX(innerBounds) + (CGFloat)0.45367 * innerBounds.size.width,
+                                 CGRectGetMinY(innerBounds) + (CGFloat)0.50083 * innerBounds.size.height)];
   [ic_clear_path
-      addLineToPoint:CGPointMake(CGRectGetMinX(innerBounds) + 0.26750f * innerBounds.size.width,
-                                 CGRectGetMinY(innerBounds) + 0.68700f * innerBounds.size.height)];
+      addLineToPoint:CGPointMake(CGRectGetMinX(innerBounds) + (CGFloat)0.26750 * innerBounds.size.width,
+                                 CGRectGetMinY(innerBounds) + (CGFloat)0.68700 * innerBounds.size.height)];
   [ic_clear_path
-      addLineToPoint:CGPointMake(CGRectGetMinX(innerBounds) + 0.31467f * innerBounds.size.width,
-                                 CGRectGetMinY(innerBounds) + 0.73417f * innerBounds.size.height)];
+      addLineToPoint:CGPointMake(CGRectGetMinX(innerBounds) + (CGFloat)0.31467 * innerBounds.size.width,
+                                 CGRectGetMinY(innerBounds) + (CGFloat)0.73417 * innerBounds.size.height)];
   [ic_clear_path
-      addLineToPoint:CGPointMake(CGRectGetMinX(innerBounds) + 0.50083f * innerBounds.size.width,
-                                 CGRectGetMinY(innerBounds) + 0.54800f * innerBounds.size.height)];
+      addLineToPoint:CGPointMake(CGRectGetMinX(innerBounds) + (CGFloat)0.50083 * innerBounds.size.width,
+                                 CGRectGetMinY(innerBounds) + (CGFloat)0.54800 * innerBounds.size.height)];
   [ic_clear_path
-      addLineToPoint:CGPointMake(CGRectGetMinX(innerBounds) + 0.68700f * innerBounds.size.width,
-                                 CGRectGetMinY(innerBounds) + 0.73417f * innerBounds.size.height)];
+      addLineToPoint:CGPointMake(CGRectGetMinX(innerBounds) + (CGFloat)0.68700 * innerBounds.size.width,
+                                 CGRectGetMinY(innerBounds) + (CGFloat)0.73417 * innerBounds.size.height)];
   [ic_clear_path
-      addLineToPoint:CGPointMake(CGRectGetMinX(innerBounds) + 0.73417f * innerBounds.size.width,
-                                 CGRectGetMinY(innerBounds) + 0.68700f * innerBounds.size.height)];
+      addLineToPoint:CGPointMake(CGRectGetMinX(innerBounds) + (CGFloat)0.73417 * innerBounds.size.width,
+                                 CGRectGetMinY(innerBounds) + (CGFloat)0.68700 * innerBounds.size.height)];
   [ic_clear_path
-      addLineToPoint:CGPointMake(CGRectGetMinX(innerBounds) + 0.54800f * innerBounds.size.width,
-                                 CGRectGetMinY(innerBounds) + 0.50083f * innerBounds.size.height)];
+      addLineToPoint:CGPointMake(CGRectGetMinX(innerBounds) + (CGFloat)0.54800 * innerBounds.size.width,
+                                 CGRectGetMinY(innerBounds) + (CGFloat)0.50083 * innerBounds.size.height)];
   [ic_clear_path
-      addLineToPoint:CGPointMake(CGRectGetMinX(innerBounds) + 0.73417f * innerBounds.size.width,
-                                 CGRectGetMinY(innerBounds) + 0.31467f * innerBounds.size.height)];
+      addLineToPoint:CGPointMake(CGRectGetMinX(innerBounds) + (CGFloat)0.73417 * innerBounds.size.width,
+                                 CGRectGetMinY(innerBounds) + (CGFloat)0.31467 * innerBounds.size.height)];
   [ic_clear_path closePath];
 
   return ic_clear_path;
